[#1] added .env config for base urls; refactoring hooks to separate files for api calls;
