+++ .env.production
@@ -1,0 +1,1 @@
+REACT_APP_BASE_URL=https://fools-compass-svc.herokuapp.com
